Noizu.AdvancedScaffolding.Helpers (Noizu.AdvancedScaffolding v1.0.1)

Functions

__default_get_context__expand_refs__(conn, params, get_context_provider, options)

Prepares a list of paths to expand keyed to a root, plus options depth. Example expand-refs=.posts.entity-image.image,.user:5 Internally paths are converted to regex strings + max depth constraints and compared to input in json methods. Regex.match?(~"^image.entity-image.user.*$", path)

__ensure_resp_content_type__(conn, content_type)

Set content-type if not set.

__send_resp__(conn, default_status, default_content_type, body)

Send API Response.

__update_expand_options__(entity, options)

Update Entity.expand! option depth and path.

__update_options__(entity, options)

Update Poison Format/Expansion options.

__update_options__(entity, context, options)

Update Poison Format/Expansion options.

__update_options__json_format__(entity, options)

Update Poison Format/Expansion options.

api_response(conn, response, context, options \\ [])

Prepare and return api response. Expand refs, strip pii, apply specific json formatting etc. to result set before returning,

banner_text(header, msg, len \\ 120, pad \\ 0)

Prepare banner string output.

expand_concurrency(options)

Extract concurrency settings from options

expand_ref?(options)

Should entity be expanded for poison/entity expand.

expand_ref?(path, depth, options \\ nil)

force_put(entity, p, v)

injects maps in path if not already populated. @example force_put(%{}, [:a, :b, :c], 1) %{a: %{b: %{c: 1}}}

format_to_atom(v, default)

Convert string json format specifier to atom if exists or default.

get_ip(conn)

get caller's IP address.

json_library()

Return Json Provider.

page(page, query)

Return query page (e.g. run next(query) until reaching desired page.