Noizu.AdvancedScaffolding.Sphinx.Type.NullableInteger (Noizu.AdvancedScaffolding v1.0.1)

Encode/Decode Nullable Integer value for Sphinx Database.

Functions

__sphinx_bits__(field, indexing, settings)

Integer bitwidth for (Sub)field. If we break a DomainObject.Entity field into multiple sub fields for sphinx we may inspect the per subfield indexing attribute here. For example if a field expands to subfields in our sphinx_expand_field method we may inject a :sub attribute in that method specifying the name of each specific subfield. [{:"#{field}_my_sub_field", @handler, put_in(indexing, [:sub], :my_sub_field)}]

__sphinx_default__(field, indexing, settings)

Default value for (Sub)field. If we break a DomainObject.Entity field into multiple sub fields for sphinx we may inspect the per subfield indexing attribute here. For example if a field expands to subfields in our sphinx_expand_field method we may inject a :sub attribute in that method specifying the name of each specific subfield. [{:"#{field}_my_sub_field", @handler, put_in(indexing, [:sub], :my_sub_field)}]

__sphinx_encoded__(field, entity, indexing, settings)

Cast (sub)field to format understood by sphinx. E.g. convert Enum atoms to their numeric value. Replace refs/entities with their universal numeric id If we break a DomainObject.Entity field into multiple sub fields for sphinx we may inspect the per subfield indexing attribute here. For example if a field expands to subfields in our sphinx_expand_field method we may inject a :sub attribute in that method specifying the name of each specific subfield. [{:"#{field}_my_sub_field", @handler, put_in(indexing, [:sub], :my_sub_field)}]

__sphinx_encoding__(field, indexing, settings)

(Sub)field sphinx field type: :attrbigint, :attruint, :attr_multi64, etc. If we break a DomainObject.Entity field into multiple sub fields for sphinx we may inspect the per subfield indexing attribute here. For example if a field expands to subfields in our __sphinx_expand_field method we may inject a :sub attribute in that method specifying the name of each specific subfield. [{:"#{field}_my_sub_field", @handler, put_in(indexing, [:sub], :my_sub_field)}]

__sphinx_expand_field__(field, indexing, settings)

Expand field from DomainObject.Entity into the list of index fields needed by sphinx. E.g. best_friend_name: %{first: f, last: l} could map to [best_friend_name_first, best_friend_name_last, best_friend_name_full] to allow searching our index for best friend by first name, last name or full name.

__sphinx_field__()

is this module a sphinx_field handler?

__sphinx_has_default__(field, indexing, settings)

Does this (sub)field have a default value?. If we break a DomainObject.Entity field into multiple sub fields for sphinx we may inspect the per subfield indexing attribute here. For example if a field expands to subfields in our sphinx_expand_field method we may inject a :sub attribute in that method specifying the name of each specific subfield. [{:"#{field}_my_sub_field", @handler, put_in(indexing, [:sub], :my_sub_field)}]

cast(v)

Casts to Ref.

cast!(value)

Same as cast/1 but raises Ecto.CastError on invalid arguments.

embed_as(_)

Callback implementation for Ecto.Type.embed_as/1.

equal?(term1, term2)

Callback implementation for Ecto.Type.equal?/2.

load(v)

Callback implementation for Ecto.Type.load/1.