Noizu.DomainObject (Noizu.AdvancedScaffolding v1.0.1)

Functions

__using__(options \\ nil)
(macro)

Setup Base Domain Object, this struct will in turn hold Entity, Repo, Index, etc. @example defmodule User do use Noizu.DomainObject Noizu.DomainObject.noizu_entity() do public_field :name end end

noizu_entity(options \\ [], list)
(macro)

Initialize a DomainObject.Entity. Caller passes in identifier and field definitions which are in turn used to generate the domain object entity's configuration options and defstruct statement. @example defmodule User do use Noizu.DomainObject Noizu.DomainObject.noizu_entity() do public_field :name end end

noizu_index(options \\ [], list)
(macro)

noizu_repo(options \\ [], list)
(macro)

noizu_scaffolding_schema(options \\ [], list)
(macro)

noizu_sphinx_handler(options \\ [])
(macro)

noizu_table(options \\ [])
(macro)

noizu_type_handler(options \\ [])
(macro)