Noizu.DomainObject.CacheHandler.Redis (Noizu.AdvancedScaffolding v1.2.23)

The Noizu.DomainObject.CacheHandler.Redis module implements the Noizu.DomainObject.CacheHandler behaviour, providing cache handling functionality using Redis.

This module is responsible for managing cache operations such as getting, setting, and deleting cache entries for DomainObjects. It uses Redis as the underlying cache storage mechanism.

Functions

cache_key/4: Generates a cache key for a given DomainObject, ref, context, and options.

delete_cache/4: Deletes a cache entry for a given DomainObject, ref, context, and options.

pre_cache/4: Pre-caches a DomainObject with a given ref, context, and options.

get_cache/4: Retrieves a cached DomainObject with a given ref, context, and options.

Code Review

⚠️ Ensure proper error handling for cache operations.

Functions

__cache_miss__(m, redis, cache_key, ref, context, options)

@TODO - Telemetry

cache_key(m, ref, context, options)

Specs

cache_key(module(), any(), any(), Keyword.t()) :: {:ok, any()} | {:error, any()}

Generates a cache key based on the given DomainObject, ref, context, and options.

Params

m: The DomainObject module.

ref: The reference to the DomainObject.

context: The request context.

options: A keyword list of options.

Returns

: The generated cache key.

: An error and reason for the failure.

delete_cache(m, ref, context, options)

Specs

delete_cache(module(), any(), any(), Keyword.t()) ::
  :ok | {:error, any()} | any()

Deletes a cache entry for the given DomainObject, ref, context, and options.

Params

m: The DomainObject module.

ref: The reference to the DomainObject.

context: The request context.

options: A keyword list of options.

Returns

:ok: The cache entry was deleted successfully.

: An error occurred due to configuration issues.

error: Any other error that may occur during cache deletion.

get_cache(m, ref, context, options)

Specs

get_cache(module(), any(), any(), Keyword.t()) :: any() | any()

Retrieves a cached DomainObject with the given ref, context, and options.

Params

m: The DomainObject module.

ref: The reference to the DomainObject.

context: The request context.

options: A keyword list of options.

Returns

The cached DomainObject.

error: An error that may occur during cache retrieval.

pre_cache(m, ref, context, options)

Specs

pre_cache(module(), any(), any(), Keyword.t()) :: any() | any()

Pre-caches a DomainObject with the given ref, context, and options.

Params

m: The DomainObject module.

ref: The reference to the DomainObject.

context: The request context.

options: A keyword list of options.

Returns

ref: The reference to the pre-cached DomainObject.

error: An error that may occur during pre-caching the DomainObject.