Noizu.DomainObject.EncodedPath.TypeHandler (Noizu.AdvancedScaffolding v1.2.23)

Functions

__sphinx_bits__(field, indexing, settings)

Integer bitwidth for (Sub)field. If we break a DomainObject.Entity field into multiple sub fields for sphinx we may inspect the per subfield indexing attribute here. For example if a field expands to subfields in our sphinx_expand_field method we may inject a :sub attribute in that method specifying the name of each specific subfield. [{:"#{field}_my_sub_field", @handler, put_in(indexing, [:sub], :my_sub_field)}]

__sphinx_default__(field, indexing, settings)

Default value for (Sub)field. If we break a DomainObject.Entity field into multiple sub fields for sphinx we may inspect the per subfield indexing attribute here. For example if a field expands to subfields in our sphinx_expand_field method we may inject a :sub attribute in that method specifying the name of each specific subfield. [{:"#{field}_my_sub_field", @handler, put_in(indexing, [:sub], :my_sub_field)}]

__sphinx_encoded__(field, entity, indexing, settings)

Cast (sub)field to format understood by sphinx. E.g. convert Enum atoms to their numeric value. Replace refs/entities with their universal numeric id If we break a DomainObject.Entity field into multiple sub fields for sphinx we may inspect the per subfield indexing attribute here. For example if a field expands to subfields in our sphinx_expand_field method we may inject a :sub attribute in that method specifying the name of each specific subfield. [{:"#{field}_my_sub_field", @handler, put_in(indexing, [:sub], :my_sub_field)}]

__sphinx_encoding__(field, indexing, settings)

(Sub)field sphinx field type: :attrbigint, :attruint, :attr_multi_64, etc. If we break a DomainObject.Entity field into multiple sub fields for sphinx we may inspect the per subfield indexing attribute here. For example if a field expands to subfields in our __sphinx_expand_field method we may inject a :sub attribute in that method specifying the name of each specific subfield. [{:"#{field}_my_sub_field", @handler, put_in(indexing, [:sub], :my_sub_field)}]

__sphinx_expand_field__(field, indexing, settings)

Expand field from DomainObject.Entity into the list of index fields needed by sphinx. E.g. best_friend_name: %{first: f, last: l} could map to [best_friend_name_first, best_friend_name_last, best_friend_name_full] to allow searching our index for best friend by first name, last name or full name.

__sphinx_field__()

is this module a sphinx_field handler?

__sphinx_has_default__(field, indexing, settings)

Does this (sub)field have a default value?. If we break a DomainObject.Entity field into multiple sub fields for sphinx we may inspect the per subfield indexing attribute here. For example if a field expands to subfields in our sphinx_expand_field method we may inject a :sub attribute in that method specifying the name of each specific subfield. [{:"#{field}_my_sub_field", @handler, put_in(indexing, [:sub], :my_sub_field)}]

__strip_inspect__(field, value, opts)

Format instance of type for Inspect output. - strip fields, simplify object, etc.

cast(field, record, type, layer, context, options)

Cast database record fields back into entity.field of this type. For example to reconstruct a TimeStamp entity by grabbing the RDMS table's created_on, modified_on, deleted_on fields.

compare(a, b)

Determine if two instances of this type match. Possibly ignoring erroneous fields such as time stamps/transient fields, etc.

dump(field, segment, value, type, arg5, context, options)

Cast database record fields to database format.

from_json(format, field, json, context, options)

Callback implementation for Noizu.DomainObject.TypeHandler.Behaviour.from_json/5.

has_query_permission?(index, field, filter, context, options)

Callback implementation for Noizu.DomainObject.SearchIndexHandler.Behaviour.has_query_permission?/5.

post_create_callback(field, entity, context, options)

post create callback. Post entity creation callback . If, for example, we wish to save this field to its own tables after creating the records for the entity containing this field

post_create_callback!(field, entity, context, options)

(Dirty) post create callback. Post entity creation callback . If, for example, we wish to save this field to its own tables after creating the records for the entity containing this field

post_delete_callback(field, entity, context, options)

post delete callback for entity fields of this type.

post_delete_callback!(field, entity, context, options)

(Dirty) post delete callback for entity fields of this type.

post_get_callback(field, entity, context, options)

Post get callback, to allow, for example, injecting data into a transient field or running additional queries needed to populate a domain object entity field.

post_get_callback!(field, entity, context, options)

(Dirty) Post get callback, to allow, for example, injecting data into a transient field or running additional queries needed to populate a domain object entity field.

post_update_callback(field, entity, context, options)

post update callback for entity fields of this type.

post_update_callback!(field, entity, context, options)

(Dirty) post update callback for entity fields of this type.

pre_create_callback(field, entity, context, options)

pre create callback. Users may have provided a simplified representation of this type, that we need to expand into the correct type and or possibly expand and write to the database before proceeding to create the entity that contains this field.

pre_create_callback!(field, entity, context, options)

(Dirty) pre create callback. Users may have provided a simplified representation of this type, that we need to expand into the correct type and or possibly expand and write to the database before proceeding to create the entity that contains this field.

pre_delete_callback(field, entity, context, options)

pre delete callback for entity fields of this type.

pre_delete_callback!(field, entity, context, options)

(Dirty) pre delete callback for entity fields of this type.

pre_update_callback(field, entity, context, options)

pre update callback for entity fields of this type.

pre_update_callback!(field, entity, context, options)

(Dirty) pre update callback for entity fields of this type.

sync(existing, update, context, options \\ nil)

Merge existing and updated instance of type.

sync!(existing, update, context, options \\ nil)

Merge existing and updated instance of type.

to_json(json_format, as_field, value, settings, context, options)