Noizu.DomainObject.IdentifierTypeBehaviour behaviour (Noizu.AdvancedScaffolding v1.2.23)

IdentifierType Behavior

The Noizu.DomainObject.IdentifierTypeBehaviour module defines the behaviour for identifier types used in the Noizu.DomainObject framework. It provides functions for encoding, decoding, validating, and preparing regex patterns for identifier types.

Callbacks

type/0: Returns the type of the identifier.

__valid_identifier__/2: Checks if a provided value is correct for the identifier type.

__sref_section_regex__/1: Prepares a regex snippet for matching the identifier.

__id_to_string__/2: Encodes a valid identifier into a string for sref encoding.

__string_to_id__/2: Decodes a string into the identifier type.

Example

defmodule DomainObject do
  defmodule Entity do
    Noizu.DomainObject.noizu_entity() do
      identifier :atom, constraint: [:foo, :bar, :bop]
      public_field :content
    end
  end
end

Example

defmodule MyDomainObject do
defmodule Entity do
  @behaviour Noizu.DomainObject.IdentifierTypeBehaviour

  @impl true
  def type, do: :atom

  @impl true
  def __valid_identifier__(identifier, configuration) do
    # Implementation logic
  end

  @impl true
  def __sref_section_regex__(configuration) do
    # Implementation logic
  end

  @impl true
  def __id_to_string__(identifier, configuration) do
    # Implementation logic
  end

  @impl true
  def __string_to_id__(string, configuration) do
    # Implementation logic
  end
end
end

Callbacks

__id_to_string__(identifier, configuration)

Specs

__id_to_string__(identifier :: any(), configuration :: any()) ::
  {:ok, String.t()} | {:error, any()}

Encode valid identifier in string form for sref encoding.

__sref_section_regex__(configuration)

Specs

__sref_section_regex__(configuration :: any()) ::
  {:ok, String.t()} | {:error, any()}

Prepare regex snippet for matching identifier

__string_to_id__(t, configuration)

Specs

__string_to_id__(String.t(), configuration :: any()) ::
  {:ok, identifier :: any()} | {:error, any()}

Decode string into identifier type.

__valid_identifier__(identifier, configuration)

Specs

__valid_identifier__(identifier :: any(), configuration :: any()) ::
  :ok | {:error, any()}

Check if provided value is correct for identifier type.

type()

Specs

type() :: atom()