Noizu.DomainObject.SchemaInfo.Default (Noizu.AdvancedScaffolding v1.0.1)

Default Implementation for the Schema behaviour.

Functions

__noizu_info__(m)

Return Map of Scaffolding Configuration Property Details

__noizu_info__(m, property)

Scaffolding Configuration Property Details (:type, :nmid_indexes, :indexes, :enums, :tables, :sref_map, ...)

cached_filter(key, app, base, type)

cache wrapper for filter_modules/3

cached_filter(key, app, base, type, filter)

cache wrapper for filter_modules/4

filter_modules(app, base, types)

Return list of Scaffolding modules in app starting with base whose noizu_info(:type) matches the provided types filter.

filter_modules(app, base, type, filter)

return list of Modules for app under base prefix of type type that match the user provided filter method.

module_children(app, scope)

Return list of modules under app that begin with scope @example module_children(:my_app, MyApp.Store) > [MyApp.Store.Electronics, MyApp.Store.Produce, MyApp.Store.FakeMustaches]

parse_sref(mod, sref)

Parse sref and return it's ref tuple or an Noizu.DomainObject.UnsupportedModule.ref if sref is correctly formatted but no handler is found in the sref_map() set. @example parse_sref(MyApp.ScaffoldingSchema, "ref.foo-bar.1234") -> {:ref, MyApp.FooBar.Entity, 1234} parse_sref(MyApp.ScaffoldingSchema, "ref.foo-not-found.1234") -> throw "UnsupportedModule ref.foo-not-found.1234" parse_sref(MyApp.ScaffoldingSchema, "ref.bar[ref.foo-bar.1234, ref.user.noizu]") -> {:ref, MyApp.Bar.Entity, [{:ref, MyApp.FooBar.Entity, 1234}, {:ref, MyApp.User.Entity, :noizu}] parse_sref(MyApp.ScaffoldingSchema, "ref.cms{1234-1.1.3@3}") -> {:ref, MyApp.CMS.Entity, {{:article, 1234}, {:version, {1,1,3}, {:revision, 3}}}