Noizu.DomainObject.TypeHandler.Behaviour behaviour (Noizu.AdvancedScaffolding v1.2.23)

Callbacks

__strip_inspect__(any, any, any)

Specs

__strip_inspect__(any(), any(), any()) :: any()

Format instance of type for Inspect output. - strip fields, simplify object, etc.

cast(any, any, any, any, any, any)

Specs

cast(any(), any(), any(), any(), any(), any()) :: any()

Cast database record fields back into entity.field of this type. For example, to reconstruct a TimeStamp entity by grabbing the RDMS table's created_on, modified_on, deleted_on fields.

compare(any, any)

Specs

compare(any(), any()) :: :eq | :neq | :lt | :gt

Determine if two instances of this type match. Possibly ignoring erroneous fields such as time stamps/transient fields, etc.

dump(any, any, any, any, any, any, any)

Specs

dump(any(), any(), any(), any(), any(), any(), any()) :: any()

Cast database record fields to database/persistence format.

from_json(format, field, json, context, options)

Specs

from_json(
  format :: any(),
  field :: atom(),
  json :: any(),
  context :: any(),
  options :: any()
) ::
  map() | nil

post_create_callback(any, any, any, any)

Specs

post_create_callback(any(), any(), any(), any()) :: any()

Post create callback. Post entity creation callback. If, for example, we wish to save this field to its own tables after creating the records for the entity containing this field.

post_create_callback!(any, any, any, any)

Specs

post_create_callback!(any(), any(), any(), any()) :: any()

(Dirty) Post create callback. Post entity creation callback. If, for example, we wish to save this field to its own tables after creating the records for the entity containing this field.

post_delete_callback(any, any, any, any)

Specs

post_delete_callback(any(), any(), any(), any()) :: any()

Post delete callback for entity fields of this type.

post_delete_callback!(any, any, any, any)

Specs

post_delete_callback!(any(), any(), any(), any()) :: any()

(Dirty) Post delete callback for entity fields of this type.

post_get_callback(any, any, any, any)

Specs

post_get_callback(any(), any(), any(), any()) :: any()

Post get callback, to allow, for example, injecting data into a transient field or running additional queries needed to populate a domain object entity field.

post_get_callback!(any, any, any, any)

Specs

post_get_callback!(any(), any(), any(), any()) :: any()

post_update_callback(any, any, any, any)

Specs

post_update_callback(any(), any(), any(), any()) :: any()

Post update callback for entity fields of this type.

post_update_callback!(any, any, any, any)

Specs

post_update_callback!(any(), any(), any(), any()) :: any()

(Dirty) Post update callback for entity fields of this type.

pre_create_callback(any, any, any, any)

Specs

pre_create_callback(any(), any(), any(), any()) :: any()

Pre create callback. Users may have provided a simplified representation of this type, that we need to expand into the correct type and or possibly expand and write to the database before proceeding to create the entity that contains this field.

pre_create_callback!(any, any, any, any)

Specs

pre_create_callback!(any(), any(), any(), any()) :: any()

(Dirty) Pre create callback. Users may have provided a simplified representation of this type, that we need to expand into the correct type and or possibly expand and write to the database before proceeding to create the entity that contains this field.

pre_delete_callback(any, any, any, any)

Specs

pre_delete_callback(any(), any(), any(), any()) :: any()

Pre delete callback for entity fields of this type.

pre_delete_callback!(any, any, any, any)

Specs

pre_delete_callback!(any(), any(), any(), any()) :: any()

(Dirty) Pre delete callback for entity fields of this type.

pre_update_callback(any, any, any, any)

Specs

pre_update_callback(any(), any(), any(), any()) :: any()

Pre update callback for entity fields of this type.

pre_update_callback!(any, any, any, any)

Specs

pre_update_callback!(any(), any(), any(), any()) :: any()

(Dirty) Pre update callback for entity fields of this type.

sync(any, any, any)

Specs

sync(any(), any(), any()) :: any()

Merge existing and updated instance of type.

sync(any, any, any, any)

Specs

sync(any(), any(), any(), any()) :: any()

sync!(any, any, any)

Specs

sync!(any(), any(), any()) :: any()

Merge existing and updated instance of type.

sync!(any, any, any, any)

Specs

sync!(any(), any(), any(), any()) :: any()