Noizu.DomainObject (Noizu.AdvancedScaffolding v1.2.23)

Functions

__using__(options \\ nil)
(macro)

Setup Base Domain Object, this struct will in turn hold Entity, Repo, Index, etc.

See

Noizu.AdvancedScaffolding.Internal.Core.Base.Behaviour

Noizu.AdvancedScaffolding.Internal.Persistence.Base.Behaviour

Noizu.AdvancedScaffolding.Internal.EntityIndex.Base.Behaviour

Noizu.AdvancedScaffolding.Internal.Json.Base.Behaviour

Example

defmodule User do
  use Noizu.DomainObject
  Noizu.DomainObject.noizu_entity() do
    public_field :name
  end
end

noizu_entity(options \\ [], list)
(macro)

Initialize a DomainObject.Entity. Caller passes in identifier and field definitions which are in turn used to generate the domain object entity's configuration options and defstruct statement.

See

Noizu.AdvancedScaffolding.Internal.Core.Entity.Behaviour

Noizu.AdvancedScaffolding.Internal.Persistence.Entity.Behaviour

Noizu.AdvancedScaffolding.Internal.EntityIndex.Entity.Behaviour

Noizu.AdvancedScaffolding.Internal.Index.Behaviour

Noizu.AdvancedScaffolding.Internal.Json.Entity.Behaviour

Example

defmodule User do
  use Noizu.DomainObject
  Noizu.DomainObject.noizu_entity() do
    public_field :name
  end
end

noizu_repo(options \\ [], list)
(macro)

Provides scaffolding for a DomainObject.Repo module. If used with no options this behavior wll provide everything needed for basic crud. get/cache/update/delete as well as providing by default a simple repo structure %Repo{ entities: [], length: 0} which may be used to pass round specific sets of records or as an embed option for domain objects provided ta TypeHandler and optional SphinxHandler is provided.

Example

defmodule MyApp.MyDomainObject do
  ...
  defmodule Repo do
     Noizu.DomainObject.noizu_repo do
     end
  end

  defmodule Repo.TypeHandler do
    require  Noizu.DomainObject
    Noizu.DomainObject.noizu_type_handler()
  end

  def pre_create_callback(field, entity, context, options) do
      # a domain object included a Repo set of entities of type MyApp.DomainObject.Entity. From this callback we may write each of these to a 12m table  for our entity.
      super(field, entity, context, options)
  end
end

noizu_schema_info(options \\ [], list)
(macro)

Configure your DomainObject Schema module. Provides runtime compiled list of modules, sref mapping/Noizu.ERP String.t support, etc.

noizu_sphinx_handler(options \\ [])
(macro)

Similar to type handler, but responsible for casting fields to a sphinx index record. In addition the field expansion support like in our type handler behaviour it also provides default values, field type (:attr_unit, :attr_multi_64, :field), and bit width (for int fields).

noizu_table(options \\ [])
(macro)

Inject Scaffolding fields into a Ecto.Table entity.

noizu_type_handler(options \\ [])
(macro)

Type Handler Behavior. Used for casting/loading embedded fields into their persistence layer format. For example domain objects may include a TimeStamp field.

  defmodule Entity do
    @universal_identifier true
    Noizu.DomainObject.noizu_entity do
      @index true
      public_field :my_image_update, nil, Noizu.Scaffolding.V3.TimeStamp.TypeHandler
    end
  end

Where the time stamp field contains a created_on, modified_on, deleted_on field. When casting to an Ecto database the nested structure can be replaced with my_image_update_created_on ,my_image_update_modified_on ,and my_image_update_deleted_on which would match DateTime fields in our Ecto Table schema.

  defmodule Noizu.DomainObject.TimeStamp.Second do
    use Noizu.SimpleObject
    @vsn 1.0
    Noizu.SimpleObject.noizu_struct() do
      date_time_handler = Application.get_env(:noizu_advanced_scaffolding, :data_time_handler, Noizu.DomainObject.DateTime.Second.TypeHandler)
      public_field :created_on, nil, date_time_handler
      public_field :modified_on, nil, date_time_handler
      public_field :deleted_on, nil, date_time_handler
    end
   #...